requirements.txt
easyocr
manga-ocr
paddlepaddle-gpu
paddleocr
deep_translator
Pillow==10.0
wget
colorama
nest_asyncio
python-dotenv